+++ src/views/Home.js
@@ -62,7 +62,7 @@ const Home = () => {
             <div className="margin-top--2 col height-sm--13 col--md-23 col--lg-14 height-md--13 height-lg--38 background--white margin-left-md--1 margin-bottom--2 js-hover-click">
               <div className="padding-left--1 padding-right--1 padding-top--2 padding-bottom--1">
                 <div className="box__content box__content--homepage height-sm--19 height-md--19 height-md--44 padding-top-lg--17">
-                  <h2 className="tiles__title tiles__title-h2--home"><Link to="/Search"><a>Reference Search</a></Link></h2>
+                  <h2 className="tiles__title tiles__title-h2--home"><Link to="/RefSearch"><a>Reference Search</a></Link></h2>
                   <span className="image-holder hide--sm hide--md-only width-lg--12"><img src={RefSearchImage} alt="" className="no-border" /></span>
                   <p className="margin-top-lg--1">Search for a reference (VAT, Company House, UBRN) and explore the data</p>
                 </div>
